+++ auth/dream-spring-auth-base/src/main/java/com/dream/springframework/auth/base/component/OrgRespCheckAdvice.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.dream.springframework.auth.base.component;
+
+import com.dream.springframework.auth.base.service.AuthorizationService;
+import com.dream.springframework.auth.base.service.OrgPermissionService;
+import org.springframework.beans.factory.ObjectProvider;
+import org.springframework.core.MethodParameter;
+import org.springframework.http.MediaType;
+import org.springframework.http.server.ServerHttpRequest;
+import org.springframework.http.server.ServerHttpResponse;
+import org.springframework.lang.NonNull;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * Checks organization resource permission and organization based authorities and roles
+ * according to response.
+ *
+ * @author DreamJM
+ */
+@ControllerAdvice
+public class OrgRespCheckAdvice extends OrgCheckService implements ResponseBodyAdvice {
+
+    public OrgRespCheckAdvice(ObjectProvider<OrgPermissionService<?>> permissionProvider, AuthorizationService authService,
+                              HttpServletRequest request) {
+        super(permissionProvider, authService, request, true);
+    }
+
+    @Override
+    public boolean supports(@NonNull MethodParameter returnType, @NonNull Class converterType) {
+        return supports(returnType);
+    }
+
+    @Override
+    public Object beforeBodyWrite(Object body, @NonNull MethodParameter returnType, @NonNull MediaType selectedContentType,
+                                  @NonNull Class selectedConverterType, @NonNull ServerHttpRequest request,
+                                  @NonNull ServerHttpResponse response) {
+        checkAdvice(body, returnType);
+        return body;
+    }
+}
